my-email-backend/requirements.txt
Flask
Flask-Cors
Flask-Mail
python-dotenv
ipinfo